pyproject.toml
[tool.poetry]
name = "lfsir"
version = "0.1.0"
description = ""
authors = ["mohammad amin alavian <m.amin.alavian@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.9, <3.12"
pandas = "^2.1.2"
pyarrow = "^14.0.0"
pyyaml = "^6.0.1"
pydantic = "^2.4.2"
tqdm = "^4.66.1"
pyodbc = "^5.0.1"
requests = "^2.31.0"
openpyxl = "^3.1.2"


[tool.poetry.group.maintain.dependencies]
boto3 = ">=1.28.28"

[tool.poetry.group.test.dependencies]
pytest = ">=7.2"

[tool.poetry.group.examples.dependencies]
jinja2 = ">=3.1"
pdf2image = ">=1.16"

[tool.poetry.group.dev.dependencies]
black = ">=23.7.0"
ipykernel = "^6.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
